game progress
There is an alien main character who moves using the arrow keys and shoots using 'a', 'w', and 'd'. The game has a static background of a Mars-type planet. Alien spaceships move towards him on the left and right and meteors drop down from the sky. Planet sprites move through the background but do not currently do anything. The game has a main screen with a quit button, a play button, and an instructions button. The quit and play buttons work, but the instructions button is not opening up the instructions menu. There are sounds in the game for almost everything. While playing the game the more aliens and meteors that the main character shoots their score increases. The aliens begin at 50 points and the meteors begin at 1000 points. The score values increase throughout the game and the difficulty also does by making the speed faster. The main character has three lives and once those are all gone they are sent back to the menu. There is a high score tracker which can track when the high score increases from game to game, but at the moment it resets when you exit the game and open it back up.