Character graphics, player 2, and giant alien
Fixed the character sprites so that their feet touch the ground. Problem with player 2 appearing in player 1 game. Moved player 2 to the sky and gave him a spaceship. Added a giant alien who moves slowly and has a giant roar. An issue happened when the aliens increased their speed exponentially fast. Turned the aliens' speed down.