Beam and alien friends
Made the beam look larger, but it still only fires one for some reason. Added a sound for when the alien friends walk around. When the alien friends are shot a cutscene happens with an angry alien king who opens a blackhole. Use key presses to change what he is saying and enlarge the blackhole. Stopped all other things from appearing during the cutscene and after it is over it is an automatic loss.